+++ Sources/SwiftyModel/Relation.swift
@@ -7,96 +7,6 @@
 
 import Foundation
 
-public typealias HasOne<T: EntityModel> = MutualRelation<T, Relations.ToOne, Relations.Optional>
-
-public typealias BelongsTo<T: EntityModel> = MutualRelation<T, Relations.ToOne, Relations.Required>
-
-public typealias HasMany<T: EntityModel> = MutualRelation<T, Relations.ToMany, Relations.Required>
-
-public typealias HasManyNonEmpty<T: EntityModel> = MutualRelation<T, Relations.ToMany, Relations.NonEmpty<T>>
-
-public typealias ToOne<T: EntityModel> = OneWayRelation<T, Relations.ToOne, Relations.Optional>
-
-public typealias FromOne<T: EntityModel> = OneWayRelation<T, Relations.ToOne, Relations.Required>
-
-public typealias ToMany<T: EntityModel> = OneWayRelation<T, Relations.ToMany, Relations.Required>
-
-public typealias ToManyNonEmpty<T: EntityModel> = OneWayRelation<T, Relations.ToMany, Relations.NonEmpty<T>>
-
-public typealias MutualRelation<T: EntityModel, Cardinality: CardinalityProtocol, Constraint: ConstraintsProtocol> = Relation<T, Relations.Mutual, Cardinality, Constraint>
-
-public typealias OneWayRelation<T: EntityModel, Cardinality: CardinalityProtocol, Constraint: ConstraintsProtocol> = Relation<T, Relations.OneWay, Cardinality, Constraint>
-
-public protocol DirectionalityProtocol {
-    
-}
-
-public protocol CardinalityProtocol {
-    static var isToMany: Bool { get }
-}
-
-public protocol ConstraintsProtocol {
-    
-}
-
-public protocol RequiredRelation {
-    
-}
-
-public protocol OptionalRelation {
-    
-}
-
-public protocol ToManyRelationValidator: ConstraintsProtocol {
-    associatedtype Entity: EntityModel
-    
-    static func validate(models: [Entity]) throws
-    
-    static func validate(ids: [Entity.ID]) throws
-}
-
-
-public enum Relations {
-    
-    public enum OneWay: DirectionalityProtocol { }
-    
-    public enum Mutual: DirectionalityProtocol { }
-    
-    public enum ToMany: CardinalityProtocol {
-        public static var isToMany: Bool { true }
-    }
-    
-    public enum ToOne: CardinalityProtocol {
-        public static var isToMany: Bool { false }
-    }
-    
-    public enum Required: ConstraintsProtocol, RequiredRelation {
-        
-    }
-    
-    public enum Optional: ConstraintsProtocol, OptionalRelation {
-        
-    }
-    
-    public struct NonEmpty<T: EntityModel>: ConstraintsProtocol, ToManyRelationValidator {
-        public enum Errors: Error {
-            case empty
-        }
-         
-        public static func validate(models: [T]) throws {
-            guard !models.isEmpty else {
-                throw Errors.empty
-            }
-        }
-        
-        public static func validate(ids: [T.ID]) throws {
-            guard !ids.isEmpty else {
-                throw Errors.empty
-            }
-        }
-    }
-}
-
 public struct Relation<T, Directionality, Cardinality, Constraints>: Hashable where T: EntityModel,
                                                                                     Directionality: DirectionalityProtocol,
                                                                                     Cardinality: CardinalityProtocol,
@@ -143,7 +53,23 @@ public extension Relation {
     }
 }
 
-public extension Relation where Cardinality == Relations.ToOne {
+public extension Relation where Cardinality == Relations.ToOne,
+                                Constraints: OptionalRelation {
+    static func set(id: T.ID) -> Self {
+        Relation(state: .faulted([id], replace: true))
+    }
+    
+    static func set(_ entity: T) -> Self {
+        Relation(state: .entity([entity], replace: true))
+    }
+
+    static var null: Self {
+        Relation(state: .none(explicitNil: true))
+    }
+}
+
+public extension Relation where Cardinality == Relations.ToOne,
+                                Constraints: RequiredRelation {
     static func set(id: T.ID) -> Self {
         Relation(state: .faulted([id], replace: true))
     }
@@ -154,9 +80,16 @@ public extension Relation where Cardinality == Relations.ToOne {
 }
 
 public extension Relation where Cardinality == Relations.ToOne,
-                                Constraints: OptionalRelation {
-    static var null: Self {
-        Relation(state: .none(explicitNil: true))
+                                Constraints: ToOneRelationValidator,
+                                Constraints.Entity == T {
+    
+    static func set(id: T.ID) -> Self {
+        Relation(state: .faulted([id], replace: true))
+    }
+    
+    static func set(_ entity: T) throws -> Self {
+        try Constraints.validate(model: entity)
+        return Relation(state: .entity([entity], replace: true))
     }
 }
